First attempt at decomposing with AMD modules.
Also significant work on the build scripts: livereload is now automatically
injected as a script in the HTML, which makes the feature work on iOS where
I don't have plug-ins. It currently points to the server hosted by the OSX
widget, but in a future change I may adapt it to grunt watch's livereload
server.